2-Dataanalyst/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
flake8==7.1.0
fonttools==4.53.1
greenlet==3.0.3
idna==3.7
kiwisolver==1.4.5
matplotlib==3.9.1
mccabe==0.7.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pyflakes==3.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.31
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
